% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulador_qr.R
\name{plot_sim_aclareos_qr}
\alias{plot_sim_aclareos_qr}
\title{Simulador de aclareos grafico con el modelo de Quinonez Barraza y Ramirez Maldonado (2019)}
\usage{
plot_sim_aclareos_qr(x)
}
\arguments{
\item{x}{Objeto creado con la funcion \code{link{sim_aclareos_qr}}}
}
\value{
Regresa una grafica con los resultados de la simulacion tabular de la funcion \code{link{sim_aclareos_qr}}
}
\description{
La funcion permite simular aclareos de forma grafica utilizando el indice de densidad de Quinonez-Barraza y Ramirez-Maldonado (2019)
}
\seealso{
\code{\link{sim_aclareos_qr}}
}
