% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulador_reineke.R
\name{plot_sim_aclareos_reineke}
\alias{plot_sim_aclareos_reineke}
\title{Simulador de aclareos grafico con el modelo de Reineke (1933)}
\usage{
plot_sim_aclareos_reineke(x)
}
\arguments{
\item{x}{Objeto creado con la funcion \code{link{sim_aclareos_reineke}}}
}
\value{
Regresa una grafica con los resultados de la simulacion tabular de la funcion \code{link{sim_aclareos_reineke}}
}
\description{
La funcion permite simular aclareos de forma grafica utilizando el indice de densidad de Reineke (1933)
}
\seealso{
\code{\link{sim_aclareos_reineke}}
}
