% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulador_qr.R
\name{sim_aclareos_qr}
\alias{sim_aclareos_qr}
\title{Simulador de aclareos tabular con el modelo de Quinonez-Barraza y Ramirez-Maldonado (2019)}
\usage{
sim_aclareos_qr(
  IDR_max,
  N,
  DCM,
  DCMR = 25,
  B1 = -0.0532,
  No.aclareos = 3,
  lim_autoclareo = 0.7,
  lim_const = 0.4,
  lim_libre = 0.2
)
}
\arguments{
\item{IDR_max}{Valor del indice de densidad maximo del sitio}

\item{N}{Numero de arboles presente en el rodal}

\item{DCM}{Diametro cuadratico medio (cm)}

\item{DCMR}{Diametro cuadratico medio de referencia (cm)}

\item{B1}{Parametro de la pendiente de Reineke. Default -1.605}

\item{No.aclareos}{Numero de aclareos que se quieren simular. Default 3 aclareos}

\item{lim_autoclareo}{Limite inferior del autoaclareo. Default 0.7}

\item{lim_const}{Limite inferior del crecimiento constante. Default 0.4}

\item{lim_libre}{Limite inferior del crecimiento libre. Default 0.2}
}
\value{
Devuelve la simulacion de aclareos a partir de la condicion de un rodal
}
\description{
La funcion permite simular aclareos utilizando el indice de densidad de Quinonez-Barraza y Ramirez-Maldonado (2019)
}
